Refactor data fetching
